Converted the core example project to .Net Core 3.1
