+++ Source/Core.Examples.L0Tests/Application/MyApplicationService.cs
