+++ Source/Core.Examples.L0Tests/Domain/MyData.cs
