+++ Source/Core.Examples.L0Tests/IoC/MyOwnIoC.cs
