+++ Source/Core.Examples.L0Tests/TestSetup/AssemblyInitializer.cs
