--- Source/Core.Examples.MsTest/Core.Examples.MsTest.csproj
@@ -1,23 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk" ToolsVersion="15.0">
-  <PropertyGroup>
-    <TargetFramework>net472</TargetFramework>
-    <FileUpgradeFlags>
-    </FileUpgradeFlags>
-    <UpgradeBackupLocation>
-    </UpgradeBackupLocation>
-    <OldToolsVersion>2.0</OldToolsVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="..\MSTest\ExceptionAssert.cs" Link="ExceptionAssert.cs" />
-    <Compile Include="..\MSTest\MultiAssert.cs" Link="MultiAssert.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="LeanTest.Core" Version="1.1.0.156" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.0.1" />
-    <PackageReference Include="MSTest.TestAdapter" Version="1.4.0" />
-    <PackageReference Include="MSTest.TestFramework" Version="1.4.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Properties\" />
-  </ItemGroup>
-</Project>
