+++ Source/Json.Examples.MsTest/Json.Examples.MsTest.csproj
@@ -17,7 +17,4 @@
   <ItemGroup>
     <Folder Include="Properties\" />
   </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Core.Examples.MsTest\Core.Examples.MsTest.csproj" />
-  </ItemGroup>
 </Project>
