+++ Source/MSTest.Examples.MsTest/MSTest.Examples.MsTest.csproj
@@ -18,9 +18,6 @@
     <Folder Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Core.Examples.MsTest\Core.Examples.MsTest.csproj" />
-  </ItemGroup>
-  <ItemGroup>
     <Compile Update="TestData\MyData.Designer.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
